+++ src/layers/DropoutLayer.h
@@ -7,15 +7,15 @@
 
 class DropoutLayer : public Layer {
    public:
-    DropoutLayer(int output_size, float dropout_chance = 0.5);
+    DropoutLayer(int output_size, double dropout_chance = 0.5);
 
-    std::vector<float> forwardpropagate(std::vector<float> input) override;
-    void out_errors(std::vector<float> target_vector) override {
+    std::vector<double> forwardpropagate(std::vector<double> input) override;
+    void out_errors(std::vector<double> target_vector) override {
         throw std::runtime_error("DropoutLayer::out_errors() is not valid");
     };
 
    private:
-    float dropout_chance;
+    double dropout_chance;
 };
 
 #endif
